Update code to send emails to authors for internal error
